// src/app/core/consts/store-keys.enum.ts
export enum StoreKeys {
  REPORT = 'REPORT',
  TYPE_REPORT = 'TYPE_REPORT',
  TYPES_REPORTS_GROUP = 'TYPES_REPORTS_GROUP',
  REPORTS_LOCAL = 'REPORTS_GROUP',

  REPORT_TEMPLATE_CONFIGURATION = 'REPORT_TEMPLATE_CONFIGURATION',

  USER_ROLES = 'USER_ROLES',
  USER_IS_ANONYMOUS = 'USER_IS_ANONYMOUS',

  SELECTED_BASE_ID = 'SELECTED_BASE_ID',
  SELECTED_SUBBASE_ID = 'SELECTED_SUBBASE_ID',

  USER_ID = 'USER_ID',
  USER_USERNAME = 'USER_USERNAME',
  USER_DETAILS = 'USER_DETAILS',

  MY_GENERATED_REPORT = 'MY_GENERATED_REPORT',
  JOBS_BY_CATEGORY = 'JOBS_BY_CATEGORY',
  JOB_CATEGORY = 'JOB_CATEGORY',
  SELECTED_INDIVIDUALS = 'SELECTED_INDIVIDUALS',
  SELECTED_JOB = 'SELECTED_JOB',
  SELECTED_JOB_CATEGORY = 'SELECTED_JOB_CATEGORY',
  SELECTED_INDIVIDUALS_BY_GROUP = 'SELECTED_INDIVIDUALS_BY_GROUP',
  SELECTED_GROUP_ID = 'SELECTED_GROUP_ID',
  PRESELECTED_IDS = 'PRESELECTED_IDS',
  SELECTED_INDIVIDUALS_IDS_BY_AREA = 'SELECTED_INDIVIDUALS_IDS_BY_AREA',
  SELECTED_LEADER_ID_BY_AREA = 'SELECTED_LEADER_ID_BY_AREA',
  SELECTED_AREA_ID = 'SELECTED_AREA_ID',
  SELECTED_COMPETENCIES = 'SELECTED_COMPETENCIES',
  SELECTED_HRFEEDBACK = 'SELECTED_HRFEEDBACK',
}